% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COEN.R
\name{COEN}
\alias{COEN}
\title{Co-expression network construction}
\usage{
COEN(
  expr,
  reg = TFs$TF_name,
  rowSample = FALSE,
  softPower = NULL,
  networkType = "unsigned",
  TOMDenom = "min",
  RsquaredCut = 0.85,
  edgeThreshold = NULL,
  trace = FALSE
)
}
\arguments{
\item{expr}{gene expression data, it is either a matrix or a data frame.
By default, each row represents a gene, each column represents a sample.}

\item{reg}{a vector of regulators. by default, these are transcription
(co-)factors defined by three literatures/databases, namely RegNet,
TRRUST, and Marbach2016.}

\item{rowSample}{logic, if TRUE, each row represents a sample.
Otherwise, each column represents a sample. The default is FALSE.}

\item{softPower}{numeric, a soft power to achieve scale free topology.
If not provided, the parameter will be picked automatically by
\code{\link{plotSoftPower}} function from the WGCNA package.}

\item{networkType}{network type. Allowed values are (unique abbreviations
of) 'unsigned' (default), 'signed', 'signed hybrid'.
See \code{\link{adjacency}}.}

\item{TOMDenom}{a character string specifying the TOM variant to be used.
Recognized values are 'min' giving the standard TOM described in Zhang
and Horvath (2005), and 'mean' in which the min function in the
denominator is replaced by mean. The 'mean' may produce better results
but at this time should be considered experimental.}

\item{RsquaredCut}{desired minimum scale free topology fitting index R^2.
The default is 0.85.}

\item{edgeThreshold}{numeric, the threshold to remove the low weighted
edges, the default is NULL, which means no edges will be removed.}

\item{trace}{logical. To show the progress or not (default).}
}
\value{
A list of two elements. weightHi: data frame of weighted edges
between regulators and targets; and networkType: 'COEN'.
}
\description{
Constructing co-expression network using Weighted gene
co-expression network analysis (WGCNA).
}
