% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEA.R
\name{DEA}
\alias{DEA}
\title{Differential expression analysis}
\usage{
DEA(
  expr,
  pData,
  method = c("Wald_DESeq2", "LRT_DESeq2", "limma", "LRT_LM"),
  minMeanExpr = NULL,
  design,
  reduced,
  contrast,
  coef = NULL,
  name,
  fitType = c("parametric", "local", "mean"),
  sfType = c("ratio", "poscounts", "iterate"),
  betaPrior,
  minReplicatesForReplace = 7,
  useT = FALSE,
  minmu = 0.5,
  parallel = FALSE,
  BPPARAM = bpparam(),
  altHypothesis = c("greaterAbs", "lessAbs", "greater", "less"),
  listValues = c(1, -1),
  cooksCutoff,
  independentFiltering = TRUE,
  alpha = 0.1,
  filter,
  theta,
  filterFun,
  addMLE = FALSE,
  blind = FALSE,
  ndups = 1,
  spacing = 1,
  block = NULL,
  correlation,
  weights = NULL,
  proportion = 0.01,
  stdev.coef.lim = c(0.1, 4),
  trend = FALSE,
  robust = FALSE,
  winsor.tail.p = c(0.05, 0.1)
)
}
\arguments{
\item{expr}{matrix or data.frame, expression profile of a set of
genes or a set of proteins. If the \code{method = 'Wald_DESeq2' or
'LRT_DESeq2'} only non-negative integer matrix (read counts by RNA
sequencing) is accepted.}

\item{pData}{data.frame, sample phenotype data. The rows of pData must
correspond to the columns of expr.}

\item{method}{either 'Wald_DESeq2', 'LRT_DESeq2', 'limma', or 'LRT_LM' for
the differential expression analysis.
\itemize{
\item When method = 'Wald_DESeq2', the Wald test in DESeq2 package is used;
\item When method = 'LRT_DESeq2', the likelihood ratio test (LRT) in DESeq2
package is used;
\item When method = 'limma', the `ls` method and empirical Bayes method in
limma package is used to calculate moderated t-statistics and differential
p-values;
\item When method = 'LRT_LM', a likelihood ratio test is performed for each
row of `expr` to compare two linear model specified by `design` and
`reduced` arguments. In this case, the fold changes are not calculated
but set to 0.
}}

\item{minMeanExpr}{numeric, the cutoff of gene average expression for
pre-filtering. The rows of `expr` with everage expression < minMeanExpr is
removed. The higher `minMeanExpr` is, the more genes are not included for
testing.}

\item{design}{either model formula or model matrix.
For method = 'LRT_DESeq2' or 'LRT_LM', the design is the full model
formula/matrix. For method = 'limma',
and if design is a formula, the model matrix is constructed using
model.matrix(design, pData), so the name of each term in the design
formula must
be included in the column names of `pData`.}

\item{reduced}{The argument is used only when method = 'LRT_DESeq2' or
'LRT_LM', it is a reduced formula/matrix to compare against.
If the design is a model matrix, `reduced` must also be a model matrix.}

\item{contrast}{The argument is used only when method = 'LRT_DESeq2',
'Wald_DESeq2', or 'limma'. \cr
When method = 'LRT_DESeq2', or 'Wald_DESeq2', it specifies what comparison
to extract from the `DESeqDataSet` object to build a results table
(when method = 'LRT_DESeq2', this does not affect the value of `stat`,
`pvalue`, or `padj`). \cr
It can be one of following three formats:
\itemize{
\item a character vector with exactly three elements: the name of a
factor in the design formula, the name of the numerator level for
the fold change, and the name of the denominator level for the fold
change;
\item a list of 1 or 2 character vector(s): the first element specifies
the names of the fold changes for the numerator, and the second
element (optional) specifies the
names of the fold changes for the denominator. These names should be
elements
of \code{getResultsNames(design, pData)};
\item a numeric contrast vector with one element for each element in
\code{getResultsNames(design, pData)}.\cr
}
When method = 'limma', It can be one of following two formats:
\itemize{
\item a numeric matrix with rows corresponding to coefficients in design
matrix and columns containing contrasts;
\item a numeric vector if there is only one contrast. Each element of the
vector corresponds to coefficients in design matrix. This is similar to
the third format of contrast when method = 'LRT_DESeq2', or 'Wald_DESeq2'.
}}

\item{coef}{The argument is used only when method = 'limma'. (Vector of)
column number or column name specifying which coefficient or contrast of
the linear model is of interest. Default is NULL.}

\item{name}{The argument is used only when method = 'LRT_DESeq2' or
'Wald_DESeq2'. the name of the individual effect (coefficient) for
building a results table.
Use this argument rather than contrast for continuous variables, individual
effects or for individual interaction terms. The value provided to name must
be an element of \code{getResultsNames(design, pData)}.}

\item{fitType}{either 'parametric', 'local', or 'mean' for the type of
fitting of dispersions to the mean intensity. This argument is used only
when method = 'Wald_DESeq2' or 'LRT_DESeq2'. See \code{\link{DESeq}} from
DESeq2 package for more details. Default is 'parametric'.}

\item{sfType}{either 'ratio', 'poscounts', or 'iterate' for the type of size
factor estimation. This argument is used only when method = either
'Wald_DESeq2' or 'LRT_DESeq2'. See \code{\link{DESeq}} from DESeq2 package
for more details. Default is 'ratio'.}

\item{betaPrior}{This argument is used only when method = either
'Wald_DESeq2' or 'LRT_DESeq2'. See \code{\link{DESeq}} from DESeq2 package
for more details.}

\item{minReplicatesForReplace}{This argument is used only when
method = either 'Wald_DESeq2' or 'LRT_DESeq2'. See \code{\link{DESeq}}
from DESeq2 package
for more details. Default is 7.}

\item{useT}{This argument is used only when method = either
'Wald_DESeq2' or 'LRT_DESeq2'. See \code{\link{DESeq}} from DESeq2 package
for more details. Default is FALSE,}

\item{minmu}{This argument is used only when method = either
'Wald_DESeq2' or 'LRT_DESeq2'. See \code{\link{DESeq}} from DESeq2 package
for more details. Default is 0.5.}

\item{parallel}{whether computing is parallel (default is FALSE).}

\item{BPPARAM}{parameters for parallel computing (default is
\code{bpparam()}).}

\item{altHypothesis}{= c('greaterAbs', 'lessAbs', 'greater', 'less').
This argument is used only when method = either
'Wald_DESeq2' or 'LRT_DESeq2'. See \code{\link{results}} from DESeq2 package
for more details. Default is 'greaterAbs'.}

\item{listValues}{This argument is used only when method = either
'Wald_DESeq2' or 'LRT_DESeq2'. See \code{\link{results}} from DESeq2 package
for more details. Default is c(1, -1),}

\item{cooksCutoff}{theshold on Cook's distance, such that if one or more
samples for a row have a distance higher, the p-value for the row is set
to NA. This argument is used only when method = either
'Wald_DESeq2' or 'LRT_DESeq2'. See \code{\link{results}} from DESeq2 package
for more details.}

\item{independentFiltering}{logical, whether independent filtering should be
applied automatically. This argument is used only when method = either
'Wald_DESeq2' or 'LRT_DESeq2'. See \code{\link{results}} from DESeq2 package
for more details. Default is TRUE.}

\item{alpha}{the significance cutoff used for optimizing the independent
filtering.  This argument is used only when method = either
'Wald_DESeq2' or 'LRT_DESeq2'. See \code{\link{results}} from DESeq2 package
for more details. Default is 0.1,}

\item{filter}{the vector of filter statistics over which the independent
filtering is optimized. By default the mean of normalized counts is used.
This argument is used only when method = either
'Wald_DESeq2' or 'LRT_DESeq2'. See \code{\link{results}} from DESeq2 package
for more details.}

\item{theta}{the quantiles at which to assess the number of rejections from
independent filtering. This argument is used only when method = either
'Wald_DESeq2' or 'LRT_DESeq2'. See \code{\link{results}} from DESeq2 package
for more details.}

\item{filterFun}{an optional custom function for performing independent
filtering and p-value adjustment. This argument is used only when
method = either 'Wald_DESeq2' or 'LRT_DESeq2'. See \code{\link{results}}
from DESeq2 package for more details.}

\item{addMLE}{if betaPrior=TRUE was used, whether the 'unshrunken' maximum
likelihood estimates (MLE) of log2 fold change should be added as a column
to the results table. This argument is used only when method = either
'Wald_DESeq2' or 'LRT_DESeq2'. See \code{\link{results}} from DESeq2 package
for more details. Default is FALSE.}

\item{blind}{logical, whether to blind the transformation to the
experimental design. This argument is used only when method = either
'Wald_DESeq2' or 'LRT_DESeq2'. See \code{\link{vst}} from DESeq2 package for
more details. Default is FALSE, which is different from the default of vst
function.}

\item{ndups}{positive integer giving the number of times each distinct
probe is printed on each array. This argument is used only when
method = 'limma'. See \code{\link{lmFit}} from limma package for more
details. Default is 1.}

\item{spacing}{positive integer giving the spacing between duplicate
occurrences of the same probe, spacing=1 for consecutive rows.
This argument is used only
when method = 'limma'. See \code{\link{lmFit}} from limma package for
more details. Default is 1.}

\item{block}{vector or factor specifying a blocking variable on the arrays.
Has length equal to the number of arrays. Must be NULL if ndups > 2.
This argument is used only when method = 'limma'. See \code{\link{lmFit}}
from limma package for more details. Default is NULL.}

\item{correlation}{the inter-duplicate or inter-technical replicate
 correlation. The correlation value should be estimated using
the \code{\link{duplicateCorrelation}}
function. This argument is used only when method = 'limma'.
See \code{\link{lmFit}}
from limma package for more details.}

\item{weights}{non-negative precision weights. Can be a numeric matrix of
individual weights of same size as the object expression matrix,
 or a numeric
vector of array weights with length equal to ncol of the expression matrix,
or a numeric vector of gene weights with length equal to nrow of the
expression
matrix. This argument is used only when method = 'limma' or 'LRT_LM'.
See \code{\link{lmFit}} from limma package for more details.
Default is NULL.}

\item{proportion}{numeric value between 0 and 1, assumed proportion of
genes which
are differentially expressed. This argument is used only when
method = 'limma'.
See \code{\link{eBayes}} from limma package for more details.
Default is 0.01.}

\item{stdev.coef.lim}{numeric vector of length 2,
assumed lower and upper limits
for the standard deviation of log2-fold-changes for differentially expressed
genes. This argument is used only when method = 'limma'.
See \code{\link{eBayes}}
from limma package for more details. Default is c(0.1, 4).}

\item{trend}{logical, should an intensity-trend be allowed for the
prior variance?
This argument is used only when method = 'limma'. See \code{\link{eBayes}}
from limma package for more details. Default is FALSE, meaning that
the prior variance is constant.}

\item{robust}{logical, should the estimation of df.prior and var.prior be
robustified against outlier sample variances? This argument is used only
when method = 'limma'. See \code{\link{eBayes}}
from limma package for more details. Default is FALSE.}

\item{winsor.tail.p}{numeric vector of length 1 or 2, giving left and right
tail proportions of x to Winsorize. Used only when method = 'limma' and
robust=TRUE. See \code{\link{eBayes}}
from limma package for more details. Default is c(0.05,0.1)}
}
\value{
a list of 3 elements: pFC (a data frame of gene, p value, and
log2FC), fullRes (full version of differential analysis),
fit (fitted model), and normExprHigh (expression data).
}
\description{
DEA is a wraper function using DESeq2 and limma package
to perform differential expression analysis.
}
\seealso{
\code{\link{lmFit}}, \code{\link{contrasts.fit}},
\code{\link{eBayes}},  \code{\link{DESeq}}, and \code{\link{results}}
}
