% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regenrichClasses.R
\name{DeaSet}
\alias{DeaSet}
\title{DeaSet object creator}
\usage{
DeaSet(pFC, fullRes, fit)
}
\arguments{
\item{pFC}{a data frame of three columns, i.e. gene (gene names/IDs),
p (differential p values), and logFC (log2 expression fold changes).}

\item{fullRes}{a full result table from differential analysis. The
contents in this table varies depending on the differential analysis
methods.}

\item{fit}{either NULL or a linear model object from differential analysis.}
}
\value{
an object of DeaSet class.
}
\description{
This is `DeaSet` object creator function.
}
\examples{
pFC = data.frame(gene = paste0('gene', seq_len(10)),
                 p = seq(0.01, 0.3, length.out = 10),
                 logFC = rnorm(10))
fullRes = NULL
fit = NULL
dSet = DeaSet(pFC, fullRes, fit)
}
