% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R
\name{dim,TopNetwork-method}
\alias{dim,TopNetwork-method}
\title{dimention of `TopNetwork` object}
\usage{
\S4method{dim}{TopNetwork}(x)
}
\arguments{
\item{x}{a `TopNetwork` object.}
}
\value{
Dimention of regulator-target network edge table.
}
\description{
dimention of `TopNetwork` object
}
\examples{
nw = newTopNetwork()
dim(nw)
}
