% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regenrichClasses.R
\docType{class}
\name{Enrich-class}
\alias{Enrich-class}
\title{Enrich class}
\description{
The `Enrich` object is to store enrichment analysis results by either
`FET` method or `GSEA` method.
}
\section{Slots}{

\describe{
\item{\code{topResult}}{data frame. The enrichment results that pass thresholds
(default threshold is 0.05).}

\item{\code{allResult}}{data frame. The enrichment results by FET or GSEA methods.}

\item{\code{gene}}{character vector indicating the genes used for enrichment
analysis.}

\item{\code{namedScores}}{numeric vector, a vector of ranked scores (decendent),
the names of the scores are the genes to perform enrichment analysis.
Here the scores are p-value of each gene.}

\item{\code{type}}{character indicating enrichment method, either 'FET' or 'GSEA'.}
}}

