% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R
\name{head,Score-method}
\alias{head,Score-method}
\alias{tail,Score-method}
\title{head or tail of Score object}
\usage{
\S4method{head}{Score}(x, ...)

\S4method{tail}{Score}(x, ...)
}
\arguments{
\item{x}{an \code{Score} object.}

\item{...}{arguments to be passed to or from other methods.}
}
\value{
Head or tail table of Score object.
}
\description{
head or tail of Score object
}
\examples{
s = newScore(letters, seq(26), seq(26), seq(26), seq(2, 0, len = 26))
s1 = head(s)
s1

s2 = tail(s)
s2

}
