% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Lyme_GSE63085}
\alias{Lyme_GSE63085}
\title{Example RNAseq dataset [Human]}
\format{
A list of 2 elements: FPKM and sampleInfo. FPKM is the 'Fragments
Per Kilobase of transcript per Million mapped reads' data, which is a
5000 (genes) * 52 (samples) data frame. sampleInfo is the information of
samples, which is 52 (samples) * 9 (features) data frame. The full version
of FPKM table contains 23615 rows, which can be downloaded from GEO
database.
}
\source{
\href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE63085}{URL}
}
\usage{
data(Lyme_GSE63085)
}
\description{
Data from an RNA sequencing experiment on peripheral mononuclear blood
cells (PBMC) of Lyme disease patients against healthy controls.
It contains a gene expression (FPKM) table (data frame) and a
sample information table (data frame).
}
\references{
Bouquet et al. (2016) mBio 7(1): e00100-16
(\href{https://www.ncbi.nlm.nih.gov/pubmed/26873097}{PubMed})
}
\keyword{datasets}
