% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show,DeaSet-method}
\alias{show,DeaSet-method}
\alias{show,TopNetwork-method}
\alias{show,Enrich-method}
\alias{show,Score-method}
\alias{show,RegenrichSet-method}
\title{methods of generic function "show"}
\usage{
\S4method{show}{DeaSet}(object)

\S4method{show}{TopNetwork}(object)

\S4method{show}{Enrich}(object)

\S4method{show}{Score}(object)

\S4method{show}{RegenrichSet}(object)
}
\arguments{
\item{object}{one object of either \code{DeaSet}, \code{TopNetwork}, 
\code{Enrich}, \code{Score}, or \code{RegenrichSet} class.}
}
\value{
show returns an invisible original \code{object}.
}
\description{
methods of generic function "show"
}
\examples{
x = newScore(letters[1:5], 1:5, 1:5, -2:2, seq(2, 1, len = 5))
show(x)

}
