% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotOrders}
\alias{plotOrders}
\title{Compare the orders of two vectors}
\usage{
plotOrders(name1, name2)
}
\arguments{
\item{name1}{a vector with first order.}

\item{name2}{a vector with anothoer second order.}
}
\value{
A plot of comparing two orders of vectors.
}
\description{
compare the orders of two vectors
}
\examples{
a = c('a1', 'a2', 'a5', 'a4')
b = c( 'a2', 'a5', 'a7', 'a4', 'a6')
plotOrders(a, b)
}
