% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{print.Score}
\alias{print.Score}
\title{Print Score object}
\usage{
\method{print}{Score}(x, ...)
}
\arguments{
\item{x}{a Score object.}

\item{...}{optional arguments to print.}
}
\value{
print.Score returns the a \code{Score} object
}
\description{
Print Score object
}
\examples{
x = newScore(letters[1:5], 1:5, 1:5, -2:2, seq(2, 1, len = 5))
print(x)
}
