% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regFET.R
\name{regFET}
\alias{regFET}
\alias{regFET,TopNetwork-method}
\title{Enrichment analysi by Fisher's exact test.}
\usage{
regFET(object, ...)

\S4method{regFET}{TopNetwork}(
  object,
  namedScores,
  namedScoresCutoffs = 0.05,
  minSize = 5,
  maxSize = 5000,
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  qvalueCutoff = 0.2,
  regAltName = NULL,
  universe = NULL
)
}
\arguments{
\item{object}{a topNetwork object, The result returned from
\code{\link{topNet}} function.}

\item{...}{additional arguments.}

\item{namedScores}{a named numeric vector of scores,
the names of the scores are the genes to perform enrichment analysis.
And the names should be the same as in the topNetwork object.
Here the scores are p-value of each gene.}

\item{namedScoresCutoffs}{the cutoff of \code{namedScores}.}

\item{minSize}{the minimum number (default 5) of target genes.}

\item{maxSize}{the maximum number (default 5000) of target genes.}

\item{pvalueCutoff}{numeric, the cutoff for adjusted enrichment p value.
This is used for obtaining the `topResult` slot in the final `Enrich`
object. Default is 0.05.}

\item{pAdjustMethod}{p adjust method, one of 'holm', 'hochberg',
'hommel', 'bonferroni', 'BH', 'BY', 'fdr', 'none'.}

\item{qvalueCutoff}{cutoff of q-value.}

\item{regAltName}{alternative names of the regulators in the pathway.}

\item{universe}{a vector of charactors. Background target genes.}
}
\value{
An \code{enrichFET} object, consisting of 5 elements:\cr
\code{topResult}, the enrichment information of regulators that
pass minSize maxSize pvalueCutoff and qvalueCutoff cutoffs;\cr
\code{allResult}, the enrichment information of all regulators;\cr
\code{gene}, the genes pass the \code{namedScoresCutoffs}; \cr
\code{universe}, total target genes in the network;\cr
and \code{geneSets}, a list of regulators (the names of the list)
and their targets (the elements of the list).
}
\description{
Enrichment for regulators based on Fisher exact test.
}
\seealso{
\code{\link{regSEA}}
}
