% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regSEA.R
\name{regSEA}
\alias{regSEA}
\alias{regSEA,TopNetwork-method}
\title{Enrichment analysi by gene set enrichment analysis (GSEA).}
\usage{
regSEA(object, ...)

\S4method{regSEA}{TopNetwork}(
  object,
  namedScores,
  minSize = 5,
  maxSize = 5000,
  pvalueCutoff = 0.05,
  nperm = 10000,
  ...
)
}
\arguments{
\item{object}{a \code{topNetwork} object, the result returned by
\code{topNet} funciton.}

\item{...}{The rest parameters in \code{\link{fgsea}} function.}

\item{namedScores}{A named numeric vector of scores,
the names of the scores are the genes to perform enrichment analysis.
And the names should be the same as in the topNetwork object.
Here the scores are p-value of each gene.}

\item{minSize}{integer, the minimum number (default 5) of target genes.}

\item{maxSize}{integer, the maximum number (default 5000) of target genes.}

\item{pvalueCutoff}{numeric, the cutoff for adjusted enrichment p value.
This is used for obtaining the `topResult` slot in the final `Enrich`
object. Default is 0.05.}

\item{nperm}{integer, number of permutations. The minimial possible
nominal p-value is about 1/nperm. The default is 10000.}
}
\value{
a list of two elements: a table with GSEA results (see
\code{\link{fgsea}}) and a ggplot object.
}
\description{
Enrichment analysi by gene set enrichment analysis (GSEA).
}
\seealso{
\code{\link{regFET}}
}
