% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TFs}
\alias{TFs}
\title{Human gene regulators}
\format{
An object of 2-column \code{data.frame}; The first column is
ENSEMBL ID of gene regulators. The second column is gene name of gene
regulators. The row name of this data frame is identical to the
ENSEMBL ID column.
}
\usage{
data(TFs)
}
\description{
The transcrpiton factors and co-factors in humans are considered the
regulators in RegEnrich. And these regulators are obtained from
(Han et al. 2015; Marbach et al. 2016; and Liu et al. 2015).
}
\references{
Han et al. (2015) Scientific Reports, 5:11432
(\href{https://www.ncbi.nlm.nih.gov/pubmed/26066708}{PubMed}),
Liu et al. (2015) Database, bav095
(\href{https://www.ncbi.nlm.nih.gov/pubmed/26424082}{PubMed}),
Marbach et al. (2016) Nature Methods, 13(4):366-70
(\href{https://www.ncbi.nlm.nih.gov/pubmed/26424082}{PubMed}).
}
\keyword{datasets}
