% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topNet.R
\name{topNet}
\alias{topNet}
\title{Filter the top edges in the network.}
\usage{
topNet(network, percent = 5, directed = TRUE, reg = NULL)
}
\arguments{
\item{network}{A data.frame of from.node (regulators),
to.node (target genes) and weight.}

\item{percent}{Numeric. The percentage (<= 10) of top edges, default is 5,
which means top 5\% of Reg-target network edges will be retained.}

\item{directed}{The type of the network: directed (\code{directed = TRUE})
and undirected
(\code{directed = FALSE}). Default is \code{TRUE}.}

\item{reg}{The gene regulators. The default is \code{NULL}.}
}
\value{
A list of cutWeight (the cutoff edge weight),
net (the genes connected by
regulators), and validReg (valid regulators).
}
\description{
Choose the top edges in the reulator-target network.
}
