% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regenrichClasses.R
\docType{class}
\name{TopNetwork-class}
\alias{TopNetwork-class}
\title{TopNetwork class}
\description{
The `TopNetwork` object is to store either a full network (the percentage
of top edges is 100%) or a sub-network (with percentage of top edges
between 0 to 10).
}
\section{Slots}{

\describe{
\item{\code{net}}{list, in which the names of elements are regulators,
and the elements are targets of the regulators indicated by the
element name.}

\item{\code{netTable}}{data frame. A table showing regulator-target
relationship. It contains 3 columns, representing 'from.gene'
('regulators'), 'to.gene' ('targets') and 'weight', respectively.}

\item{\code{validReg}}{vector, a set of valid regulators in the network.}

\item{\code{tarReg}}{list, in which the names of elements are targets,
and the elements are regulators of the targets indicated by the
element name.}

\item{\code{directed}}{logical, whether the network is directed.}

\item{\code{networkConstruction}}{character, by which method this network
is constructed. Either 'COEN' (coexpression network using WGCNA),
or 'GRN' (gene regulatory network using random forest), or 'new'
(a network provided by the user).}

\item{\code{percent}}{numeric, what percentage of the top edges are remained.
The value must be between 0 (excluding) and 100 (including).}
}}

