% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regenrichClasses.R
\name{TopNetwork}
\alias{TopNetwork}
\title{TopNetwork object creator}
\usage{
TopNetwork(
  networkEdgeTable,
  reg = TFs$TF_name,
  directed = TRUE,
  networkConstruction = c("new", "COEN", "GRN"),
  percent = 100,
  extendObject = FALSE
)
}
\arguments{
\item{networkEdgeTable}{a data frame of 3 columns, representing 'from.gene'
('regulators'), 'to.gene' ('targets') and 'weight', respectively.}

\item{reg}{a vector of gene regulators.}

\item{directed}{logical, whether the network is directed. Default is TRUE.}

\item{networkConstruction}{the method to construct this network.
Possible can be:
'COEN', coexpression network;
'GRN', gene regulatory network by random forest;
'new' (default), meaning a network provided by user, rather than
infered based
on the expression data.}

\item{percent}{the percentage of edges in the original whole network.
Default is 100, meaning 100\% edges in whole network.}

\item{extendObject}{logical, whether net and tarReg will be constructed.
Default is FALSE, which can save memory. This can be set FALSE for
construting 'network' slot
in RegenrichSet objects, but must be set TRUE when construting
'topNetP' slot.}
}
\value{
an object of topNetwork class.
}
\description{
This function create `TopNetwork` object using 3-column edge table.
}
\examples{
edge = data.frame(from = rep(TFs$TF_name[seq_len(3)], seq_len(3)),
                  to = TFs$TF_name[11:16], weight = 0.1*(6:1))
object = TopNetwork(edge, networkConstruction = 'new', percent = 100)
}
