% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regenrichClasses.R
\docType{class}
\name{DeaSet-class}
\alias{DeaSet-class}
\title{DeaSet class}
\value{
an object of DeaSet class.
}
\description{
The `DeaSet` object is to store differnetial expression analysis
results, for further enrichment and regulator ranking.
}
\section{Slots}{

\describe{
\item{\code{pFC}}{a data frame of three columns, i.e. gene (gene names/IDs),
p (differential p values), and logFC (log2 expression fold changes).}

\item{\code{fullRes}}{a full result table from differential analysis. The
contents in this table varies depending on the differential analysis
methods.}

\item{\code{fit}}{either NULL or a linear model object from differential analysis.}
}}

