% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotSoftPower}
\alias{plotSoftPower}
\title{Plot soft power for WGCNA analysis}
\usage{
plotSoftPower(
  expr,
  rowSample = FALSE,
  powerVector = c(seq_len(10), seq(12, 20, by = 2)),
  RsquaredCut = 0.85,
  networkType = "unsigned",
  verbose = 0,
  ...
)
}
\arguments{
\item{expr}{Gene expression data, either a matrix or a data frame.
By default, each row represents a gene, each column represents a sample.}

\item{rowSample}{logic. If \code{TRUE}, each row represents a sample.
The default is \code{FALSE}.}

\item{powerVector}{a vector of soft thresholding powers for which the scale
free topology fit indices are to be calculated.}

\item{RsquaredCut}{desired minimum scale free topology fitting index R^2.
The default is 0.85.}

\item{networkType}{character, network type. Allowed values are
(unique abbreviations of) "unsigned" (default), "signed", "signed hybrid".
See \code{\link{adjacency}}.}

\item{verbose}{integer level of verbosity. 0 (default) means silent, higher
values make the output progressively more and more verbose.}

\item{...}{Parameters in \code{\link{pickSoftThreshold}} function.}
}
\value{
a list of three elements: \code{powerEstimate}, \code{fitIndices},
and \code{plot}.
The details about \code{powerEstimate} and \code{fitIndices} are shown
in \code{\link{pickSoftThreshold}}. The \code{plot} is a ggplot object.
}
\description{
Plot soft power and corresponding scale free topology
fitting index to find a proper soft power for WGCNA analysis.
}
\examples{
data(Lyme_GSE63085)
log2FPKM = log2(Lyme_GSE63085$FPKM + 1)
log2FPKMhi = log2FPKM[rowMeans(log2FPKM) >= 10^-3, , drop = FALSE]
log2FPKMhi = head(log2FPKMhi, 3000) # First 3000 genes for example
softP = plotSoftPower(log2FPKMhi, RsquaredCut = 0.85)
}
\seealso{
\code{\link{pickSoftThreshold}}
}
