% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regenrichClasses.R
\docType{class}
\name{TopNetwork-class}
\alias{TopNetwork-class}
\title{TopNetwork class}
\description{
The `TopNetwork` object is to store either a full network (the percentage
of top edges is 100%) or a sub-network (with percentage of top edges
between 0 to 10).
}
\section{Slots}{

\describe{
\item{\code{element}}{tibble, the pool of targets in the network.}

\item{\code{set}}{tibble, the pool of valid regulators.}

\item{\code{elementset}}{tibble, regulator-target edges with edge weights.
and the elements are regulators of the targets indicated by the
element name.}

\item{\code{directed}}{logical, whether the network is directed.}

\item{\code{networkConstruction}}{character, by which method this network
is constructed. Either 'COEN' (coexpression network using WGCNA),
or 'GRN' (gene regulatory network using random forest), or 'new'
(a network provided by the user).}

\item{\code{percent}}{numeric, what percentage of the top edges are remained.
The value must be between 0 (excluding) and 100 (including).}

\item{\code{active}}{character, which data table is activated, the default is 
"elementset".}
}}

